pyproject.toml
[build-system]
requires = ["setuptools>=70"]
build-backend = "setuptools.build_meta"

[project]
name = "aip-intern"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.4",
    "langchain-openai>=0.3",
    "langchain-mcp-adapters>=0.1",
    "mcp>=1.0",
    "crewai>=0.80",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "langfuse>=2.0",
    "pandas>=2.0",
    "matplotlib>=3.0",
    "ipython>=8.0",
    "notebook>=7.0",
    "nest-asyncio>=1.6",
    "python-dotenv>=1.0",
    "tiktoken>=0.7",
    "litellm>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.4",
    "nbstripout>=0.7",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
markers = ["integration: requires live LLM endpoint (deselect with -m 'not integration')"]
testpaths = ["tests"]

[tool.ruff.lint]
select = ["E", "F", "I"]
